+++ test/unit/fuzz-tools.test.js
@@ -274,6 +274,7 @@ import * as fastDementiaV294 from '../../lib/fast-dementia-v294.js';
 import * as gdsV295 from '../../lib/gds-v295.js';
 import * as benzoEquivV296 from '../../lib/benzo-equiv-v296.js';
 import * as nerveInjuryV297 from '../../lib/nerve-injury-v297.js';
+import * as concussionRtsV298 from '../../lib/concussion-rts-v298.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -493,6 +494,7 @@ const MODULES = {
   'gds-v295.js': gdsV295,
   'benzo-equiv-v296.js': benzoEquivV296,
   'nerve-injury-v297.js': nerveInjuryV297,
+  'concussion-rts-v298.js': concussionRtsV298,
   'tb-testing.js': tbTesting,
 };